WIP - more debug statements
