+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/internal/multipart/MultipartDownloadResumeContext.java
@@ -71,6 +71,7 @@ public class MultipartDownloadResumeContext {
     }
 
     public void addCompletedPart(int partNumber) {
+        System.out.println("Completed part: " + partNumber);
         completedParts.add(partNumber);
     }
